Allow URL-encoded spaces in generated HTML
After generating the HTML with xmlgen, the HTML is passed to `format` as a
format string. However, “%” is not an uncommon character in HTML, most commonly
with “%20” for URL-encoded spaces. `format` would complain when it encountered
these.

This is a workaround to specifically get `format` to ignore `%20`. A more
general solution would be better, but this fixes my immediate issue of entries
having spaces in their names.

This also wraps the “Error during parsing of entry text” message in a `<p>`.